refactor: created new component
