+++ apps/commudle-admin/src/app/feature-modules/events/components/event-locations/event-location-tracks/event-location-track/event-location-track.component.html
@@ -1,0 +1,55 @@
+<!-- <nb-accordion *ngFor="let elt of eventLocationTracks; let elti = index" class="com-mb-3 com-shadow-none">
+  <nb-accordion-item
+    collapsed="{{ !trackSlotVisibility[elt.id] }}"
+    (click)="changeTrackSlotVisibility(!trackSlotVisibility[elt.id], elt.id)"
+  >
+    <nb-accordion-item-header class="com-justify-between">
+      <p class="track-name">{{ elt.name }}</p>
+      &nbsp;
+      <nb-icon *ngIf="elt.embedded_video_stream" icon="video-outline"></nb-icon>
+      <div *ngIf="event.editable" class="com-mr-8">
+        <button (click)="showEditTrackForm(elt)" nbButton size="tiny" status="info">
+          <nb-icon icon="edit-outline"></nb-icon>
+        </button>
+        <button (click)="confirmDeleteTrack(elt)" nbButton size="tiny" status="danger">
+          <nb-icon icon="close"></nb-icon>
+        </button> -->
+<!-- <button (click)="showAddSlotForm(elt)" nbButton size="tiny" status="primary">
+          <nb-icon icon="plus"></nb-icon>
+        </button> -->
+<!-- </div>
+    </nb-accordion-item-header>
+    <nb-accordion-item-body class="com-rounded-b-2xl">
+      <nb-list *ngIf="trackSlotVisibility[elt.id]">
+        <nb-list-item *ngFor="let slot of sortedTrackSlots[elt.id]; let tsi = index">
+          <!-- *ngFor="let slot of sortedTrackSlots[elt.id] elt.track_slots; let tsi = index" -->
+<!-- <div class="session-time">
+            <p>{{ moment(slot.start_time).format('HH:mm') }}</p>
+            <p>{{ moment(slot.end_time).format('HH:mm') }}</p>
+          </div>
+          <div class="session-details">
+            <div class="session-title">
+              <div *ngIf="slot.tags_list">
+                <app-badge *ngFor="let tag of slot.tags_list.split(',')" [fontSize]="'small'" [text]="tag"></app-badge>
+              </div>
+              <p>{{ slot.session_title }}</p>
+              <div *ngIf="slot.user" class="session-speaker">
+                <app-user-profile-card-small [user]="slot.user"></app-user-profile-card-small>
+              </div>
+            </div>
+          </div>
+          <button (click)="showEditSlotForm(slot); $event.stopPropagation()" nbButton size="tiny" status="info">
+            <nb-icon icon="edit-outline"></nb-icon>
+          </button>
+          <button (click)="confirmDeleteSlot(slot); $event.stopPropagation()" nbButton size="tiny" status="danger">
+            <nb-icon icon="close"></nb-icon>
+          </button> -->
+<!-- <button (click)="addSlot()" nbButton size="tiny" status="primary">
+            <nb-icon icon="plus-circle-outline"></nb-icon>
+    </button> -->
+<!-- <div *ngIf="elt.track_slots.length === 0">no slots yet</div>
+        </nb-list-item>
+      </nb-list>
+    </nb-accordion-item-body>
+  </nb-accordion-item>
+</nb-accordion> -->
