--- apps/shared-components/event-location-track/event-location-track.component.html
@@ -1,1 +1,0 @@
-<p>event-location-track works!</p>
