--- apps/shared-components/event-location-track/event-location-track.component.ts
@@ -1,12 +1,0 @@
-import { Component, OnInit } from '@angular/core';
-
-@Component({
-  selector: 'commudle-event-location-track',
-  templateUrl: './event-location-track.component.html',
-  styleUrls: ['./event-location-track.component.scss'],
-})
-export class EventLocationTrackComponent implements OnInit {
-  constructor() {}
-
-  ngOnInit(): void {}
-}
